services/websockets/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
websockets==13.1